refactor: split SessionSettings component into smaller units
- Extract EditSessionDialog component (with confirm-restart mode)
- Extract RestartSessionDialog component
- Extract CloseSessionDialog component
- Extract DeleteWorktreeDialog component
- Extract SessionSettingsMenu component
- Reduce SessionSettings.tsx from 518 lines to 84 lines
- Use shadcn/ui Dialog and AlertDialog components

Closes #43
