+++ packages/client/src/components/session-settings/CloseSessionDialog.tsx
@@ -1,0 +1,77 @@
+import { useState } from 'react';
+import { useNavigate } from '@tanstack/react-router';
+import {
+  AlertDialog,
+  AlertDialogContent,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogAction,
+  AlertDialogCancel,
+} from '../ui/alert-dialog';
+import { deleteSession } from '../../lib/api';
+
+export interface CloseSessionDialogProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  sessionId: string;
+}
+
+export function CloseSessionDialog({
+  open,
+  onOpenChange,
+  sessionId,
+}: CloseSessionDialogProps) {
+  const navigate = useNavigate();
+  const [isSubmitting, setIsSubmitting] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const handleCloseSession = async () => {
+    setIsSubmitting(true);
+    setError(null);
+
+    try {
+      await deleteSession(sessionId);
+      onOpenChange(false);
+      navigate({ to: '/' });
+    } catch (err) {
+      setError(err instanceof Error ? err.message : 'Failed to close session');
+    } finally {
+      setIsSubmitting(false);
+    }
+  };
+
+  const handleClose = () => {
+    setError(null);
+    onOpenChange(false);
+  };
+
+  return (
+    <AlertDialog open={open} onOpenChange={handleClose}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Close Session</AlertDialogTitle>
+          <AlertDialogDescription asChild>
+            <div className="space-y-2">
+              <p>Are you sure you want to close this session?</p>
+              <p className="text-xs text-gray-500">
+                This will stop the Claude process. The worktree will remain and
+                you can start a new session from it later.
+              </p>
+            </div>
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        {error && <p className="text-sm text-red-400">{error}</p>}
+        <AlertDialogFooter>
+          <AlertDialogCancel disabled={isSubmitting}>
+            Cancel
+          </AlertDialogCancel>
+          <AlertDialogAction onClick={handleCloseSession} disabled={isSubmitting}>
+            {isSubmitting ? 'Closing...' : 'Close'}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
