+++ packages/client/src/components/session-settings/RestartSessionDialog.tsx
@@ -1,0 +1,91 @@
+import { useState } from 'react';
+import {
+  AlertDialog,
+  AlertDialogContent,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogAction,
+  AlertDialogCancel,
+} from '../ui/alert-dialog';
+import { restartAgentWorker, getSession } from '../../lib/api';
+
+export interface RestartSessionDialogProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  sessionId: string;
+  onSessionRestart?: () => void;
+}
+
+export function RestartSessionDialog({
+  open,
+  onOpenChange,
+  sessionId,
+  onSessionRestart,
+}: RestartSessionDialogProps) {
+  const [isSubmitting, setIsSubmitting] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const handleRestart = async (continueConversation: boolean) => {
+    setIsSubmitting(true);
+    setError(null);
+
+    try {
+      // Get the session to find the first agent worker
+      const session = await getSession(sessionId);
+      if (!session) {
+        throw new Error('Session not found');
+      }
+      const agentWorker = session.workers.find(w => w.type === 'agent');
+      if (!agentWorker) {
+        throw new Error('No agent worker found');
+      }
+      await restartAgentWorker(sessionId, agentWorker.id, continueConversation);
+      onOpenChange(false);
+      if (onSessionRestart) {
+        onSessionRestart();
+      }
+    } catch (err) {
+      setError(
+        err instanceof Error ? err.message : 'Failed to restart session'
+      );
+    } finally {
+      setIsSubmitting(false);
+    }
+  };
+
+  const handleClose = () => {
+    setError(null);
+    onOpenChange(false);
+  };
+
+  return (
+    <AlertDialog open={open} onOpenChange={handleClose}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Restart Session</AlertDialogTitle>
+          <AlertDialogDescription>
+            How would you like to restart this session?
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        {error && <p className="text-sm text-red-400">{error}</p>}
+        <AlertDialogFooter>
+          <AlertDialogCancel disabled={isSubmitting}>
+            Cancel
+          </AlertDialogCancel>
+          <button
+            onClick={() => handleRestart(false)}
+            className="btn bg-slate-600 hover:bg-slate-500"
+            disabled={isSubmitting}
+          >
+            New Session
+          </button>
+          <AlertDialogAction onClick={() => handleRestart(true)} disabled={isSubmitting}>
+            {isSubmitting ? 'Restarting...' : 'Continue (-c)'}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
