+++ packages/client/src/components/session-settings/index.ts
@@ -1,0 +1,14 @@
+export { EditSessionDialog } from './EditSessionDialog';
+export type { EditSessionDialogProps } from './EditSessionDialog';
+
+export { RestartSessionDialog } from './RestartSessionDialog';
+export type { RestartSessionDialogProps } from './RestartSessionDialog';
+
+export { CloseSessionDialog } from './CloseSessionDialog';
+export type { CloseSessionDialogProps } from './CloseSessionDialog';
+
+export { DeleteWorktreeDialog } from './DeleteWorktreeDialog';
+export type { DeleteWorktreeDialogProps } from './DeleteWorktreeDialog';
+
+export { SessionSettingsMenu } from './SessionSettingsMenu';
+export type { SessionSettingsMenuProps, MenuAction } from './SessionSettingsMenu';
